Dealing with at least once messaging part 2
We can pass commitId to command , or the builder can give it to you.
Event builder will generate a commit Id everytime an event is created.
Version  number is part of every command and event. For command it it
the last known version, for event it is the current version of the
aggregate.